+++ XObjectsTests/CodeGenerationTests.cs
@@ -1,8 +1,12 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Text.RegularExpressions;
 using System.Xml;
+using System.Xml.Resolvers;
+using LinqToXsd;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -13,6 +17,8 @@ using Xml.Schema.Linq.CodeGen;
 
 namespace Xml.Schema.Linq.Tests
 {
+    using SF = SyntaxFactory;
+
     public class CodeGenerationTests
     {
         [Test]
@@ -33,6 +39,89 @@ namespace Xml.Schema.Linq.Tests
             var cSharpNsName = Regex.Replace(namespaceNode.Name.ToString(), @"\W", "_");
 
             Assert.IsTrue(cSharpNsName == nsName);
+        }
+
+        /// <summary>
+        /// Tests the the BuildWrapperDictionary() method of the LinqToXsdTypeManager class that's
+        /// generated does not contain <c>typeof(void)</c> expressions, which are meaningless and break
+        /// typed XElement conversion.
+        /// </summary>
+        [Test]
+        public void AtomNoVoidTypeOfExpressionsInLinqToXsdTypeManagerBuildWrapperDictionaryMethodTest()
+        {
+            const string atomDir = @"Schemas\Atom";
+            var atomXsdFolder = Path.Combine(Environment.CurrentDirectory, atomDir);
+            var atomRssXsdFile = $"{atomXsdFolder}\\atom.xsd";
+            var atomRssXsdFileInfo = new FileInfo(atomRssXsdFile);
+            var tree = Utilities.GenerateSyntaxTree(atomRssXsdFileInfo);
+
+            var linqToXsdTypeManagerClassDeclarationSyntax = tree.GetNamespaceRoot().DescendantNodes().OfType<ClassDeclarationSyntax>()
+                                                                 .FirstOrDefault(cds => cds.Identifier.ValueText == nameof(LinqToXsdTypeManager));
+
+            Assert.IsNotNull(linqToXsdTypeManagerClassDeclarationSyntax);
+
+            var buildWrapperDictionaryMethod = linqToXsdTypeManagerClassDeclarationSyntax
+                                               .DescendantNodes().OfType<MethodDeclarationSyntax>()
+                                               .FirstOrDefault(mds =>
+                                                   mds.Identifier.ValueText == "BuildWrapperDictionary");
+
+            Assert.IsNotNull(buildWrapperDictionaryMethod);
+
+            var statements = buildWrapperDictionaryMethod.DescendantNodes().OfType<InvocationExpressionSyntax>().ToArray();
+
+            Assert.IsTrue(statements.Length == 2);
+
+            var typeOfExpressions = statements.SelectMany(ies => ies.ArgumentList.DescendantNodes()).OfType<TypeOfExpressionSyntax>().ToArray();
+
+            Assert.IsNotEmpty(typeOfExpressions);
+            Assert.IsTrue(typeOfExpressions.Length == 4);
+
+            var typeOfVoid = SF.ParseExpression("typeof(void)");
+            var nonVoidTypeOfExpressions = typeOfExpressions.Where(toe => !toe.IsEquivalentTo(typeOfVoid)).ToArray();
+            var voidTypeOfExpressions = typeOfExpressions.Except(nonVoidTypeOfExpressions).ToArray();
+
+            Assert.IsNotEmpty(nonVoidTypeOfExpressions);
+            Assert.IsTrue(nonVoidTypeOfExpressions.Length == 4);
+
+            // if this is not empty, then you have a problem...
+            Assert.IsEmpty(voidTypeOfExpressions);
+        }
+
+        /// <summary>
+        /// There shouldn't be <c>typeof(void)</c> expressions in any generated code.
+        /// <para>See commit bc75ea0 which introduced this incorrect behaviour.</para>
+        /// </summary>
+        [Test]
+        public void NoVoidTypeOfExpressionsInGeneratedCodeEver()
+        {
+            var dir = new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, "Schemas"));
+            var allXsds = dir.GetFiles("*.xsd", SearchOption.AllDirectories)
+                // Microsoft.Build schemas will have typeof(void) expressions due to the existence of bugs that predate this .net core port
+                .Where(f => !f.FullName.Contains("Microsoft.Build.")) 
+                .Select(f => f.FullName).ToArray();
+
+            var allProcessableXsds = FileSystemUtilities.ResolvePossibleFileAndFolderPathsToProcessableSchemas(allXsds)
+                .Select(fp => new FileInfo(fp));
+
+            foreach (var xsd in allProcessableXsds) {
+                var generatedCodeTree = Utilities.GenerateSyntaxTree(xsd);
+
+                var root = generatedCodeTree.GetRoot();
+
+                var allDescendents = root.DescendantNodes().SelectMany(d => d.DescendantNodes());
+                var allStatements = allDescendents.OfType<StatementSyntax>();
+                var allExpressions = allStatements.SelectMany(s => s.DescendantNodes()).OfType<ExpressionSyntax>();
+                var typeOfExpressions = allExpressions.OfType<TypeOfExpressionSyntax>().Distinct().ToArray();
+
+                Assert.IsNotEmpty(typeOfExpressions);
+
+                var typeOfVoid = SF.ParseExpression("typeof(void)");
+                var nonVoidTypeOfExpressions = typeOfExpressions.Where(toe => !toe.IsEquivalentTo(typeOfVoid)).ToArray();
+                var voidTypeOfExpressions = typeOfExpressions.Where(toe => toe.IsEquivalentTo(typeOfVoid)).ToArray();
+
+                Assert.IsNotEmpty(nonVoidTypeOfExpressions);
+                Assert.IsEmpty(voidTypeOfExpressions);
+            }
         }
     }
 }