+++ RELEASENOTES.md
@@ -1,5 +1,21 @@
 # LinqToXsdCore Release Notes
 
+## LinqToXsdCore 3.0.0.12 and XObjectsCore 3.0.0.11
+Nuget packages:
+* https://www.nuget.org/packages/LinqToXsdCore/3.0.0.12
+* https://www.nuget.org/packages/XObjectsCore/3.0.0.11
+	* When a group of XSD files or a folder of them all import or include each other, LinqToXsd cannot decide which one to use as the entry point for code generation, so now the CLI throws an exception when that condition is met while trying to resolve which XSD file to use.
+	* Reverts "Avoid type name conflicts in generated code" from previous release, as it broke the code generation of the `BuildWrapperDictionary()` method generated inside the `LinqToXsdTypeManager`; it adds `typeof(void)` expressions, which breaks untyped `XElement` type conversion. Previous (and correct) behavior was to add `typeof(T)` expressions where T was the generated complex or global element type.
+	* Fixes an issue whereby setting a string value to an attribute whose type was `AnyAtomicType` resulted in an error.
+	* Fixes an issue when using the static Parse() or Load() methods on an internal generated type.
+
+## LinqToXsdCore 3.0.0.11 and XObjectsCore 3.0.0.10
+Nuget packages:
+* https://www.nuget.org/packages/LinqToXsdCore/3.0.0.11
+* https://www.nuget.org/packages/XObjectsCore/3.0.0.10
+	* Avoid type name conflicts in generated code. 
+	* Do not prefix an identifier with the '@' character when not needed.
+
 ## XObjectsCore 3.0.0.9
 Nuget packages: 
 * https://www.nuget.org/packages/XObjectsCore/3.0.0.9
@@ -9,7 +25,7 @@ Added `XTypedElementEqualityComparer` and `XTypedElementDeepEqualityComparer` cl
 ## LinqToXsdCore 3.0.0.10 and XObjectsCore 3.0.0.8
 Nuget packages: 
 * https://www.nuget.org/packages/XObjectsCore/3.0.0.8
-* https://www.nuget.org/packages/LinqToXsdCore/3.0.0.9
+* https://www.nuget.org/packages/LinqToXsdCore/3.0.0.10
 
 Modified the behaviour of retrieving the value of an attribute, when the schema type is anyAtomicType (which is the default for attributes when no type is given). The value literal is now returned as a string (pre-existing behaviour would throw an exception saying that anyAtomicType is not a supported conversion to the CLR type 'string').